Update how-to-access-folder-contents.md
- Full guides for Windows, ChomeOS and basic GUI for Linux. 

Todo

- MacOS needs screenshots (I don't have access to a mac atm)
- Screenshots could be better for Windows and Chrome, I will do new ones when I can (these are from a video from my brother)

**Mobile**

Need to do guides for android and iOS, I don't know enough about it atm. I can do a quick one with research, or you can if that's easier. I have left a "coming soon" at the bottom as I don't want it to seem like this isn't possible (I know it can be done)